# tests/server: support bundle binary