@vszakats 2025-03-12
\o/ first successful run with the bundled server binary:
https://github.com/curl/curl/actions/runs/13822611417/job/38671280902
@ https://github.com/curl/curl/pull/15000/commits/fd9aff1e9c1cfb0552c168a3fa472cd4b2434a08

The, or at least the final, hurdle was runtime confusion (possibly server crash?) due to building servers with `CURLDEBUG` and/or `DEBUGBUILD` in bundle mode. Fixed it with a series of corrections. Regardless, an option may be nice to build the servers without those (as default even).