@vszakats 2025-03-15
I plan to merge this tomorrow if no objection (and if I finish in time
with the performance gain table.)

It's extending the existing test bundle feature, that is disabled by
default. The patch touches the non-bundle build path minimally, for
the test targets.