@vszakats 2025-03-17
Here are two CI run with and without test bundles enabled:

without: https://github.com/curl/curl/actions/runs/13909934284
with: https://github.com/curl/curl/actions/runs/13909397643

Some examples:

Cygwin AM: 10m1 vs 1m32 (6.5x)
Cygwin CM: 1m27 vs 28s
mingw, CM clang-x86_64 gnutls: 1m4 vs 6s
dl-mingw, CM R Ninja: 3m33 vs 1m36
dl-mingw, CM R Make: 1m47 vs 25
dl-mingw, CM D Make !unity: 2m41 vs 30s
linux-mingw CM: 1m2s 10s
MSVC CM: 3m32 vs 11s (20x)

For MSVC it's a 20x reduction, others are also pretty good.